Refactor: Move GetThreatModelUseCase to scan > details > usecases
